Move wccom redirect to theme step of profiler
